+++ generator/lib/src/element.dart
@@ -20,7 +20,6 @@ import 'dart:math';
 
 import 'package:analyzer/dart/analysis/results.dart';
 import 'package:analyzer/dart/ast/ast.dart';
-import 'package:analyzer/dart/constant/value.dart';
 import 'package:analyzer/dart/element/element.dart';
 import 'package:realm_generator/src/annotation_value.dart';
 import 'package:realm_generator/src/expanded_context_span.dart';