+++ lib/src/realm_object.dart
@@ -169,6 +169,8 @@ class RealmCoreAccessor implements RealmAccessor {
               return object.realm.createList<RealmObject>(handle, listMetadata);
             case ObjectType.embeddedObject:
               return object.realm.createList<EmbeddedObject>(handle, listMetadata);
+            case ObjectType.asymmetricObject:
+              return object.realm.createList<AsymmetricObject>(handle, listMetadata);
             default:
               throw RealmError('List of ${listMetadata.schema.baseType} is not supported yet');
           }
@@ -310,6 +312,7 @@ mixin RealmObjectBase on RealmEntity implements RealmObjectBaseMarker, Finalizab
     _typeOf<RealmObject?>(): () => _ConcreteRealmObject(),
     EmbeddedObject: () => _ConcreteEmbeddedObject(),
     _typeOf<EmbeddedObject?>(): () => _ConcreteEmbeddedObject(),
+    // We can never read asymmetric objects from a realm, so we don't need a factory for them.
   };
 
   /// @nodoc
@@ -455,6 +458,17 @@ mixin RealmObject on RealmObjectBase implements RealmObjectMarker {}
 
 /// @nodoc
 mixin EmbeddedObject on RealmObjectBase implements EmbeddedObjectMarker {}
+
+/// Base for any object that can be persisted in a [Realm], but cannot be retrieved,
+/// hence cannot be modified.
+///
+/// The benefit of using [AsymmetricObject] is that synchronization is one-way, and
+/// thus performs much better. However, they cannot be queried, or retrieved
+/// locally, which limits their use-cases greatly.
+///
+/// Use [AsymmetricObject] when you have a write-/only use case. You use it by
+/// parsing [ObjectType.asymmetricObject] to the [RealmModel] annotation.
+mixin AsymmetricObject on RealmObjectBase implements AsymmetricObjectMarker {}
 
 extension EmbeddedObjectExtension on EmbeddedObject {
   /// Retrieve the [parent] object of this embedded object.
@@ -684,6 +698,8 @@ class DynamicRealmObject {
     _typeOf<RealmObject?>(): RealmPropertyType.object,
     EmbeddedObject: RealmPropertyType.object,
     _typeOf<EmbeddedObject?>(): RealmPropertyType.object,
+    AsymmetricObject: RealmPropertyType.object,
+    _typeOf<AsymmetricObject?>(): RealmPropertyType.object,
   };
 
   RealmPropertyType? _getPropertyType<T extends Object?>() => _propertyTypeMap[T];