+++ CHANGELOG.md
@@ -11,6 +11,7 @@
    * `SyncClientErrorCode`, `SyncConnectionErrorCode`, `SyncSessionErrorCode`, `SyncWebSocketErrorCode`, `GeneralSyncErrorCode, SyncErrorCategory`
 * Added new Sync errors types `BadFlexibleSyncQueryError`, `WrongSyncTypeError` and `UnrecoverableSyncError`. ([#1387](https://github.com/realm/realm-dart/pull/1387))
 * Throw an exception if `File::unlock` has failed, in order to inform the SDK that we are likely hitting some limitation on the OS filesystem, instead of crashing  the application and use the same file locking logic for all the platforms. (Core upgrade)
+* Support asymmetric objects. ([#1400](https://github.com/realm/realm-dart/pull/1400))
 
 ### Fixed
 * Crash when querying the size of a Object property through a link chain (Core upgrade, since v13.17.2)